refactor: split routes into modules
