+++ server/routes/index.ts
@@ -2,347 +2,18 @@ import type { Express } from "express";
 import { createServer, type Server } from "http";
 import { setupAuth, requireAdmin } from "./auth";
 import { storage } from "./storage";
-import { ContentService } from "./content-service";
 import { SiteCode, SITE_CODES } from "@shared/schema";
-
+import contentRoutes from "./content";
+import jobsRoutes from "./jobs";
+import newsRoutes from "./news";
 export async function registerRoutes(app: Express): Promise<Server> {
   // Setup authentication routes and middleware
   setupAuth(app);
 
-  // Initialize content service
-  const contentService = new ContentService(storage);
+  app.use("/api", contentRoutes);
+  app.use("/api", jobsRoutes);
+  app.use("/api", newsRoutes);
 
-  // API Routes
-  // =====================================
-
-  // Content Routes
-  app.get("/api/content/:pageId", async (req, res) => {
-    try {
-      const { pageId } = req.params;
-      const lang = req.query.lang?.toString() || "pt";
-      const site = req.query.site?.toString();
-      
-      // Verifica se o site é válido
-      const siteCode = site && (SITE_CODES as readonly string[]).includes(site) 
-        ? site as SiteCode 
-        : undefined;
-      
-      const content = await contentService.getContent(pageId, lang, siteCode);
-      
-      if (!content) {
-        return res.status(404).json({ message: "Content not found" });
-      }
-      
-      res.json(content);
-    } catch (error) {
-      console.error("Error fetching content:", error);
-      res.status(500).json({ message: "Error fetching content" });
-    }
-  });
-
-  app.put("/api/content/:pageId", async (req, res) => {
-    if (!req.isAuthenticated()) {
-      return res.status(401).json({ message: "Unauthorized" });
-    }
-
-    try {
-      const { pageId } = req.params;
-      const lang = req.query.lang?.toString() || "pt";
-      const site = req.query.site?.toString();
-      const content = req.body;
-
-      // Verifica se o site é válido
-      const siteCode = site && (SITE_CODES as readonly string[]).includes(site) 
-        ? site as SiteCode 
-        : undefined;
-
-      const result = await contentService.updateContent(pageId, lang, content, siteCode);
-      
-      // Log activity
-      await storage.createActivityLog({
-        userId: req.user!.id,
-        action: "update",
-        entityType: "content",
-        entityId: pageId,
-        details: { language: lang, site: siteCode }
-      });
-      
-      res.json(result);
-    } catch (error) {
-      console.error("Error updating content:", error);
-      res.status(500).json({ message: "Error updating content" });
-    }
-  });
-
-  // Jobs Routes
-  app.get("/api/jobs", async (req, res) => {
-    try {
-      const lang = req.query.lang?.toString() || "pt";
-      const site = req.query.site?.toString();
-      
-      // Verifica se o site é válido
-      const siteCode = site && (SITE_CODES as readonly string[]).includes(site) 
-        ? site as SiteCode 
-        : undefined;
-      
-      const jobs = await storage.getJobs(lang, siteCode);
-      res.json(jobs);
-    } catch (error) {
-      console.error("Error fetching jobs:", error);
-      res.status(500).json({ message: "Error fetching jobs" });
-    }
-  });
-
-  app.get("/api/jobs/featured", async (req, res) => {
-    try {
-      const lang = req.query.lang?.toString() || "pt";
-      const site = req.query.site?.toString();
-      
-      // Verifica se o site é válido
-      const siteCode = site && (SITE_CODES as readonly string[]).includes(site) 
-        ? site as SiteCode 
-        : undefined;
-      
-      const jobs = await storage.getFeaturedJobs(lang, siteCode);
-      res.json(jobs);
-    } catch (error) {
-      console.error("Error fetching featured jobs:", error);
-      res.status(500).json({ message: "Error fetching featured jobs" });
-    }
-  });
-
-  app.get("/api/jobs/:id", async (req, res) => {
-    try {
-      const { id } = req.params;
-      const lang = req.query.lang?.toString() || "pt";
-      const job = await storage.getJob(parseInt(id), lang);
-      
-      if (!job) {
-        return res.status(404).json({ message: "Job not found" });
-      }
-      
-      res.json(job);
-    } catch (error) {
-      res.status(500).json({ message: "Error fetching job" });
-    }
-  });
-
-  app.post("/api/jobs", async (req, res) => {
-    if (!req.isAuthenticated()) {
-      return res.status(401).json({ message: "Unauthorized" });
-    }
-
-    try {
-      const job = req.body;
-      const lang = req.query.lang?.toString() || "pt";
-      
-      const result = await storage.createJob({
-        ...job,
-        language: lang
-      });
-      
-      // Log activity
-      await storage.createActivityLog({
-        userId: req.user!.id,
-        action: "create",
-        entityType: "job",
-        entityId: result.id.toString(),
-        details: { title: result.title }
-      });
-      
-      res.status(201).json(result);
-    } catch (error) {
-      res.status(500).json({ message: "Error creating job" });
-    }
-  });
-
-  app.put("/api/jobs/:id", async (req, res) => {
-    if (!req.isAuthenticated()) {
-      return res.status(401).json({ message: "Unauthorized" });
-    }
-
-    try {
-      const { id } = req.params;
-      const job = req.body;
-      const lang = req.query.lang?.toString() || "pt";
-      
-      const result = await storage.updateJob(parseInt(id), {
-        ...job,
-        language: lang
-      });
-      
-      // Log activity
-      await storage.createActivityLog({
-        userId: req.user!.id,
-        action: "update",
-        entityType: "job",
-        entityId: id,
-        details: { title: job.title }
-      });
-      
-      res.json(result);
-    } catch (error) {
-      res.status(500).json({ message: "Error updating job" });
-    }
-  });
-
-  app.delete("/api/jobs/:id", async (req, res) => {
-    if (!req.isAuthenticated()) {
-      return res.status(401).json({ message: "Unauthorized" });
-    }
-
-    try {
-      const { id } = req.params;
-      await storage.deleteJob(parseInt(id));
-      
-      // Log activity
-      await storage.createActivityLog({
-        userId: req.user!.id,
-        action: "delete",
-        entityType: "job",
-        entityId: id
-      });
-      
-      res.json({ success: true });
-    } catch (error) {
-      res.status(500).json({ message: "Error deleting job" });
-    }
-  });
-
-  // News Routes
-  app.get("/api/news", async (req, res) => {
-    try {
-      const lang = req.query.lang?.toString() || "pt";
-      const site = req.query.site?.toString();
-      
-      // Verifica se o site é válido
-      const siteCode = site && (SITE_CODES as readonly string[]).includes(site) 
-        ? site as SiteCode 
-        : undefined;
-      
-      const newsItems = await storage.getNewsItems(lang, siteCode);
-      res.json(newsItems);
-    } catch (error) {
-      console.error("Error fetching news:", error);
-      res.status(500).json({ message: "Error fetching news" });
-    }
-  });
-
-  app.get("/api/news/latest", async (req, res) => {
-    try {
-      const lang = req.query.lang?.toString() || "pt";
-      const site = req.query.site?.toString();
-      
-      // Verifica se o site é válido
-      const siteCode = site && (SITE_CODES as readonly string[]).includes(site) 
-        ? site as SiteCode 
-        : undefined;
-      
-      const newsItems = await storage.getLatestNews(lang, siteCode);
-      res.json(newsItems);
-    } catch (error) {
-      console.error("Error fetching latest news:", error);
-      res.status(500).json({ message: "Error fetching latest news" });
-    }
-  });
-
-  app.get("/api/news/:id", async (req, res) => {
-    try {
-      const { id } = req.params;
-      const lang = req.query.lang?.toString() || "pt";
-      const newsItem = await storage.getNewsItem(parseInt(id), lang);
-      
-      if (!newsItem) {
-        return res.status(404).json({ message: "News item not found" });
-      }
-      
-      res.json(newsItem);
-    } catch (error) {
-      res.status(500).json({ message: "Error fetching news item" });
-    }
-  });
-
-  app.post("/api/news", async (req, res) => {
-    if (!req.isAuthenticated()) {
-      return res.status(401).json({ message: "Unauthorized" });
-    }
-
-    try {
-      const newsItem = req.body;
-      const lang = req.query.lang?.toString() || "pt";
-      
-      const result = await storage.createNewsItem({
-        ...newsItem,
-        language: lang
-      });
-      
-      // Log activity
-      await storage.createActivityLog({
-        userId: req.user!.id,
-        action: "create",
-        entityType: "news",
-        entityId: result.id.toString(),
-        details: { title: result.title }
-      });
-      
-      res.status(201).json(result);
-    } catch (error) {
-      res.status(500).json({ message: "Error creating news item" });
-    }
-  });
-
-  app.put("/api/news/:id", async (req, res) => {
-    if (!req.isAuthenticated()) {
-      return res.status(401).json({ message: "Unauthorized" });
-    }
-
-    try {
-      const { id } = req.params;
-      const newsItem = req.body;
-      const lang = req.query.lang?.toString() || "pt";
-      
-      const result = await storage.updateNewsItem(parseInt(id), {
-        ...newsItem,
-        language: lang
-      });
-      
-      // Log activity
-      await storage.createActivityLog({
-        userId: req.user!.id,
-        action: "update",
-        entityType: "news",
-        entityId: id,
-        details: { title: newsItem.title }
-      });
-      
-      res.json(result);
-    } catch (error) {
-      res.status(500).json({ message: "Error updating news item" });
-    }
-  });
-
-  app.delete("/api/news/:id", async (req, res) => {
-    if (!req.isAuthenticated()) {
-      return res.status(401).json({ message: "Unauthorized" });
-    }
-
-    try {
-      const { id } = req.params;
-      await storage.deleteNewsItem(parseInt(id));
-      
-      // Log activity
-      await storage.createActivityLog({
-        userId: req.user!.id,
-        action: "delete",
-        entityType: "news",
-        entityId: id
-      });
-      
-      res.json({ success: true });
-    } catch (error) {
-      res.status(500).json({ message: "Error deleting news item" });
-    }
-  });
 
   // Translation API Route
   app.post("/api/translate", async (req, res) => {
@@ -795,179 +466,6 @@ export async function registerRoutes(app: Express): Promise<Server> {
     } catch (error) {
       console.error("Error fetching file versions:", error);
       res.status(500).json({ message: "Error fetching file versions" });
-    }
-  });
-
-  // Jobs Management Routes
-  app.get("/api/jobs/applications", async (req, res) => {
-    if (!req.isAuthenticated()) {
-      return res.status(401).json({ message: "Unauthorized" });
-    }
-
-    try {
-      // Mock applications data - in production, this would come from database
-      const applications = [
-        {
-          id: 1,
-          jobId: 1,
-          name: "Maria Silva",
-          email: "maria.silva@email.com",
-          phone: "+55 11 99999-9999",
-          coverLetter: "Tenho grande interesse na vaga de Desenvolvedor React. Possuo 3 anos de experiência com React, TypeScript e Node.js. Trabalhei em projetos de e-commerce e sistemas de gestão.",
-          resumeUrl: "/uploads/resumes/maria-silva-cv.pdf",
-          status: "pending",
-          appliedAt: "2024-01-15T10:30:00Z"
-        },
-        {
-          id: 2,
-          jobId: 1,
-          name: "João Santos",
-          email: "joao.santos@email.com",
-          phone: "+55 11 88888-8888",
-          coverLetter: "Sou desenvolvedor fullstack com foco em React e Python. Tenho experiência em desenvolvimento de APIs REST e integração com bancos de dados.",
-          resumeUrl: "/uploads/resumes/joao-santos-cv.pdf",
-          status: "reviewing",
-          appliedAt: "2024-01-14T14:20:00Z"
-        },
-        {
-          id: 3,
-          jobId: 2,
-          name: "Ana Costa",
-          email: "ana.costa@email.com",
-          phone: "+55 11 77777-7777",
-          coverLetter: "Especialista em marketing digital com 5 anos de experiência. Tenho conhecimento em SEO, Google Ads, redes sociais e análise de dados.",
-          status: "accepted",
-          appliedAt: "2024-01-12T09:15:00Z"
-        }
-      ];
-
-      res.json(applications);
-    } catch (error) {
-      console.error("Error fetching job applications:", error);
-      res.status(500).json({ message: "Error fetching job applications" });
-    }
-  });
-
-  app.post("/api/jobs/apply", async (req, res) => {
-    try {
-      const { jobId, name, email, phone, coverLetter } = req.body;
-      
-      // Validate required fields
-      if (!jobId || !name || !email || !coverLetter) {
-        return res.status(400).json({ message: "Missing required fields" });
-      }
-
-      // Create new application
-      const newApplication = {
-        id: Date.now(),
-        jobId: parseInt(jobId),
-        name,
-        email,
-        phone,
-        coverLetter,
-        status: "pending",
-        appliedAt: new Date().toISOString()
-      };
-
-      // Log activity
-      await storage.createActivityLog({
-        userId: 0, // System user for job applications
-        action: "job_application",
-        entityType: "job_application",
-        entityId: newApplication.id.toString(),
-        details: { jobId, name, email }
-      });
-
-      res.status(201).json(newApplication);
-    } catch (error) {
-      console.error("Error creating job application:", error);
-      res.status(500).json({ message: "Error creating job application" });
-    }
-  });
-
-  app.put("/api/jobs/applications/:id/status", async (req, res) => {
-    if (!req.isAuthenticated()) {
-      return res.status(401).json({ message: "Unauthorized" });
-    }
-
-    try {
-      const { id } = req.params;
-      const { status } = req.body;
-      
-      // Validate status
-      const validStatuses = ['pending', 'reviewing', 'accepted', 'rejected'];
-      if (!validStatuses.includes(status)) {
-        return res.status(400).json({ message: "Invalid status" });
-      }
-
-      // Update application status
-      const updatedApplication = {
-        id: parseInt(id),
-        status,
-        updatedAt: new Date().toISOString(),
-        updatedBy: req.user!.username
-      };
-
-      // Log activity
-      await storage.createActivityLog({
-        userId: req.user!.id,
-        action: "update",
-        entityType: "job_application",
-        entityId: id,
-        details: { status, action: "status updated" }
-      });
-
-      res.json(updatedApplication);
-    } catch (error) {
-      console.error("Error updating application status:", error);
-      res.status(500).json({ message: "Error updating application status" });
-    }
-  });
-
-  // Public job listing for candidates
-  app.get("/api/public/jobs", async (req, res) => {
-    try {
-      const lang = req.query.lang?.toString() || "pt";
-      const site = req.query.site?.toString();
-      
-      // Verifica se o site é válido
-      const siteCode = site && (SITE_CODES as readonly string[]).includes(site) 
-        ? site as SiteCode 
-        : undefined;
-      
-      const jobs = await storage.getJobs(lang, siteCode);
-      
-      // Filter only active jobs for public API
-      const activeJobs = jobs.filter(job => job.active);
-      
-      res.json(activeJobs);
-    } catch (error) {
-      console.error("Error fetching public jobs:", error);
-      res.status(500).json({ message: "Error fetching jobs" });
-    }
-  });
-
-  app.get("/api/public/jobs/:slug", async (req, res) => {
-    try {
-      const { slug } = req.params;
-      const lang = req.query.lang?.toString() || "pt";
-      const site = req.query.site?.toString();
-      
-      // Verifica se o site é válido
-      const siteCode = site && (SITE_CODES as readonly string[]).includes(site) 
-        ? site as SiteCode 
-        : undefined;
-      
-      const job = await storage.getJobBySlug(slug, lang, siteCode);
-      
-      if (!job || !job.active) {
-        return res.status(404).json({ message: "Job not found" });
-      }
-      
-      res.json(job);
-    } catch (error) {
-      console.error("Error fetching job:", error);
-      res.status(500).json({ message: "Error fetching job" });
     }
   });
